Fix exceptions in tests
