+++ fbchat/__init__.py
@@ -17,6 +17,7 @@ from ._exception import (
     HTTPError,
     ParseError,
     ExternalError,
+    GraphQLError,
     InvalidParameters,
     NotLoggedIn,
     PleaseRefresh,
