+++ tests/test_send.py
@@ -1,7 +1,7 @@
 import pytest
 
 from os import path
-from fbchat import FBchatFacebookError, Message, Mention
+from fbchat import Message, Mention
 from utils import subset, STICKER_LIST, EMOJI_LIST, TEXT_LIST
 
 pytestmark = pytest.mark.online
